Fix routing for post categories
