+++ src/routes/blog/CategoryBadge.svelte
@@ -4,6 +4,6 @@
 	export let category: Tables<'post_categories'>;
 </script>
 
-<a href={`/blog?categoria=${category.slug}`} class="badge badge-neutral hover:badge-primary"
+<a href={`/blog/categoria/${category.slug}`} class="badge badge-neutral hover:badge-primary"
 	>{category?.title}</a
 >
